refactor: extract Global IFW Rule UI components and screens
- Move shared UI components (`RuleMetaBadge`, `RulePrimaryMetadataText`, `RuleSecondaryMetadataText`, `DeleteConfirmationDialog`) from `GlobalIfwRuleScreen.kt` to a new `GlobalIfwRuleUiComponents.kt` file
- Extract `AdvancedGlobalIfwRuleDetailScreen` into its own file for better maintainability
- Update `GlobalIfwRuleScreen` to use the new common text components and refined spacing (from 8dp/12dp to 10dp) for metadata sections
- Standardize typography and colors across IFW rule displays using the new shared components
